first version on GH
This version of the workflow uses collections and runs Trimmomatic, HISAT, and Stringtie. The Gene abundance collection of Stringtie is joined to generate a complete list of expressed transcripts in a population, this collection is also cut to include only gene ID (XLOC) and FPKM values and collpsed to one file - this file is filtered to include only genes with an FPKM > 1 and expressed in 10% of the population (at FPKM > 1). The filtered list of compared to the complete list to provide a final output of expressed genes across a population of cells with corresponding FPKM values per gene per cell.