Workflows associated with v7.9
The latest iteration of the workflow uses cell based (>200000 read pairs passing QC, >%60 mapping rate, and > 400 expressed genes/cell) and the expression based (a gene needs to be expressed > 1 FPKM in 10% of the sub-population to be included) filtering to generate expression tables. This iteration also includes a "cell based report" on cell based metrics. 

The additional workflows are useful for preprocessing data ahead of clustering and heatmap visualizations.